apply changes from #687 to current master
