+++ core/src/main/java/net/sourceforge/jnlp/util/whitelist/UrlWhiteListUtils.java
@@ -2,8 +2,6 @@ package net.sourceforge.jnlp.util.whitelist;
 
 import net.adoptopenjdk.icedteaweb.Assert;
 import net.adoptopenjdk.icedteaweb.StringUtils;
-import net.adoptopenjdk.icedteaweb.logging.Logger;
-import net.adoptopenjdk.icedteaweb.logging.LoggerFactory;
 import net.sourceforge.jnlp.runtime.JNLPRuntime;
 import net.sourceforge.jnlp.util.IpUtil;
 
@@ -26,22 +24,24 @@ public class UrlWhiteListUtils {
     public static final String HTTP = "http";
     private static final String PROTOCOL_SEPARATOR = "://";
 
-    private static final Logger LOG = LoggerFactory.getLogger(UrlWhiteListUtils.class);
-
     private static List<WhitelistEntry> applicationUrlWhiteList;
     private static final Lock whiteListLock = new ReentrantLock();
 
     public static List<WhitelistEntry> getApplicationUrlWhiteList() {
-        whiteListLock.lock();
-        try {
             if (applicationUrlWhiteList == null) {
-                applicationUrlWhiteList = JNLPRuntime.getConfiguration().getPropertyAsList(KEY_SECURITY_SERVER_WHITELIST)
-                        .stream()
-                        .filter(s -> !StringUtils.isBlank(s))
-                        .map(UrlWhiteListUtils::validateWhitelistUrl)
-                        .collect(Collectors.toList());
+                applicationUrlWhiteList = whitelistPropertyName(KEY_SECURITY_SERVER_WHITELIST);
             }
             return applicationUrlWhiteList;
+    }
+
+    public static List<WhitelistEntry> whitelistPropertyName(String whitelistPropertyName) {
+        whiteListLock.lock();
+        try {
+        return JNLPRuntime.getConfiguration().getPropertyAsList(whitelistPropertyName)
+                .stream()
+                .filter(s -> !StringUtils.isBlank(s))
+                .map(UrlWhiteListUtils::validateWhitelistUrl)
+                .collect(Collectors.toList());
         } finally {
             whiteListLock.unlock();
         }
@@ -51,7 +51,7 @@ public class UrlWhiteListUtils {
         return isUrlInWhitelist(url, getApplicationUrlWhiteList());
     }
 
-    static boolean isUrlInWhitelist(final URL url, final List<WhitelistEntry> whiteList) {
+    public static boolean isUrlInWhitelist(final URL url, final List<WhitelistEntry> whiteList) {
         Assert.requireNonNull(url, "url");
         Assert.requireNonNull(whiteList, "whiteList");
